these are the most recent working modules.
You need all three of these files in the same file directory for it to work. Only execute the parent class--the other two classes are imported. We need to add the Mims ENDOR processing functions into the initialization of the MimsENDOR class. See comments in that file about which ones to add.